requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#    pip install -r requirements.txt

#



-i https://pypi.org/simple

bcrypt==3.2.0; python_version >= '3.6'

blinker==1.4

cffi==1.14.5

click==8.0.1; python_version >= '3.6'

colorama==0.4.4; platform_system == 'Windows'

django==2.2.5

dnspython==2.1.0; python_version >= '3.6'

email-validator==1.1.3

flask-bcrypt==0.7.1

flask-login==0.5.0

flask-mail==0.9.1

flask-sqlalchemy==2.5.1

flask-wtf==0.15.1

flask==2.0.1

greenlet==1.1.0; python_version >= '3'

gunicorn==19.9.0

idna==3.2; python_version >= '3.5'

image==1.5.33

importlib-metadata==4.5.0; python_version < '3.8'

itsdangerous==2.0.1; python_version >= '3.6'

jinja2==3.0.1; python_version >= '3.6'

joblib==1.0.1; python_version >= '3.6'

markupsafe==2.0.1; python_version >= '3.6'

numpy==1.20.3

pandas==1.2.4

pillow==8.2.0

pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

pytz==2019.2

scikit-learn==0.24.2

scipy==1.6.3; python_version < '3.10' and python_version >= '3.7'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlalchemy==1.4.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

sqlparse==0.3.0

threadpoolctl==2.1.0; python_version >= '3.5'

typing-extensions==3.10.0.0; python_version < '3.8'

virtualenv==16.7.5

virtualenvwrapper-win==1.2.5

werkzeug==2.0.1; python_version >= '3.6'

wtforms==2.3.3

xgboost==1.3.3

zipp==3.4.1; python_version >= '3.6'

